requirements.txt
# Telegram
telethon==1.34.0
aiogram==3.4.1

# AI
httpx==0.25.2

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.25

# Utils
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
structlog==24.1.0
